# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BrewMaster AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Add a new equipment item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Remove equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Sensors ────────────────────────────────────────────────────────────────
...
  /actuators:
    get:
      operationId: listActuators
      tags: [actuators]
      summary: List all actuators
      responses:
        "200":
          description: Actuators list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actuator"
    post:
      operationId: createActuator
      tags: [actuators]
      summary: Add a new actuator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actuator"
  /actuators/{id}/toggle:
    post:
      operationId: toggleActuator
      tags: [actuators]
      summary: Toggle an actuator on/off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated actuator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actuator"

  # ─── Alarms ─────────────────────────────────────────────────────────────────
  /alarms:
    get:
      operationId: listAlarms
      tags: [alarms]
      summary: List alarms
      responses:
        "200":
          description: Alarms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alarm"
    post:
      operationId: createAlarm
      tags: [alarms]
      summary: Create an alarm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlarmInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alarm"
...
  /alarms/{id}/acknowledge:
    post:
      operationId: acknowledgeAlarm
      tags: [alarms]
      summary: Acknowledge an alarm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Acknowledged alarm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alarm"

  # ─── History ────────────────────────────────────────────────────────────────
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send message to AI brewing assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"
  /ai/diagnose:
    post:
      operationId: aiDiagnose
      tags: [ai]
      summary: Run AI diagnostic on current production
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiDiagnoseInput"
      responses:
        "200":
          description: Diagnosis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDiagnosisResult"

  # ─── Simulator ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Equipment:
      type: object
      required: [id, name, type, status, connected, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [fermenter, kettle, mash_tun, hlt, pump, valve, chiller, co2_tank, other]
        status:
          type: string
          enum: [operational, maintenance, offline, error]
        connected:
          type: boolean
        firmwareVersion:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    EquipmentInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [fermenter, kettle, mash_tun, hlt, pump, valve, chiller, co2_tank, other]
        firmwareVersion:
          type: string
        location:
          type: string
        notes:
          type: string
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [fermenter, kettle, mash_tun, hlt, pump, valve, chiller, co2_tank, other]
        status:
          type: string
          enum: [operational, maintenance, offline, error]
        connected:
          type: boolean
        firmwareVersion:
          type: string
        location:
          type: string
        notes:
          type: string

    # ─── Sensors ───────────────────────────────────────────────────────────────
...
    Actuator:
      type: object
      required: [id, name, type, isOn, equipmentId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [heater, pump, valve, agitator, cooler, fan, light]
        isOn:
          type: boolean
        equipmentId:
          type: integer
        equipmentName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ActuatorInput:
      type: object
      required: [name, type, equipmentId]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [heater, pump, valve, agitator, cooler, fan, light]
        equipmentId:
          type: integer
        notes:
          type: string

    # ─── Alarms ────────────────────────────────────────────────────────────────
    Alarm:
      type: object
      required: [id, message, priority, severity, acknowledged, triggeredAt, createdAt]
      properties:
        id:
          type: integer
        message:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        severity:
          type: string
          enum: [info, warning, error, emergency]
        acknowledged:
          type: boolean
        acknowledgedAt:
          type: ["string", "null"]
        equipmentId:
          type: ["integer", "null"]
        equipmentName:
          type: ["string", "null"]
        sensorId:
          type: ["integer", "null"]
        triggeredAt:
          type: string
        createdAt:
          type: string
    AlarmInput:
      type: object
      required: [message, priority, severity]
      properties:
        message:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        severity:
          type: string
          enum: [info, warning, error, emergency]
        equipmentId:
          type: integer
        sensorId:
          type: integer

    # ─── Events ────────────────────────────────────────────────────────────────
...
    AiMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        context:
          type: ["string", "null"]
    AiMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
    AiDiagnoseInput:
      type: object
      required: [productionId]
      properties:
        productionId:
          type: integer
    AiDiagnosisResult:
      type: object
      required: [diagnosis, suggestions, confidence]
      properties:
        diagnosis:
          type: string
        suggestions:
          type: array
          items:
            type: string
        confidence:
          type: number
        warnings:
          type: array
          items:
            type: string