# lib/api-zod/src/generated/types/aiMessageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BrewMaster AI Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiMessageInput {
  message: string;
  /** @nullable */
  context?: string | null;
}