# lib/api-zod/src/generated/types/alarmInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BrewMaster AI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AlarmInputPriority } from './alarmInputPriority';
import type { AlarmInputSeverity } from './alarmInputSeverity';

export interface AlarmInput {
  message: string;
  priority: AlarmInputPriority;
  severity: AlarmInputSeverity;
  equipmentId?: number;
  sensorId?: number;
}